added view-controller model
