+++ package.json
@@ -5,7 +5,8 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start-dev": "nodemon server.js"
+    "start": "node ./src/server.js",
+    "start-dev": "nodemon ./src/server.js"
   },
   "keywords": [],
   "author": "",
